Functracer: fix verification for repeated return addresses
When addresses are from Thumb instructions, they have bit0 set,
breakpoint_from_address() expect to always receive an address with bit0
no set, because it is how it is stored in the struct breakpoint
"address" field. This patch fixes this issue, which caused broken
behaviour for realloc() calls in Thumb code.

Signed-off-by: Anderson Lizardo <anderson.lizardo@indt.org.br>
Signed-off-by: Bruna Moreira <bruna.moreira@indt.org.br>