Functracer: replace traditional breakpoint approach with SSOL
This commit replaces the traditional breakpoint management mechanism (inherited
from ltrace) with a new approach based on the "single-step out of line" idea,
as (fairly well) described in
http://sourceware.org/ml/systemtap/2007-q1/msg00571.html. Whereas in this
message the technique was designed to be implemented in kernel code, the basic
concept could be implemented in user space.

In short, SSOL consists of allowing to pass over breakpoints without having to
temporarily disable them to execute the original instruction. This is possible
because the original instruction is stored in a injected mmaped area in the
target process memory space (here called "SSOL area"). Whenever a breakpoint is
hit, the PC/EIP is redirected to the original instruction stored in this area,
single-stepped, and then redirected to the next instruction.

One of the side-effects of this approach is that, for breakpoints that are
never disabled, it is not required to temporarily disable them, which on
traditional breakpoint scheme would allow for a concurrent thread to pass over
the breakpoint without hitting it. SSOL-based breakpoints are never disabled,
thus all threads will catch them.

This approach works well for function entry breakpoints, but for return
breakpoints it would require many entries (one per return address) and
managing them would be complex. It was implemented a novel approach
that consists on modifying the return address of the function call on
function entry, to point to a unique address that contains a breakpoint
instruction. This allows for having a single breakpoint for every
function return, and the call stack is managed by a already existing
structure in functracer.

Signed-off-by: Anderson Lizardo <anderson.lizardo@indt.org.br>
Signed-off-by: Bruna Moreira <bruna.moreira@indt.org.br>