Functracer: fix SSOL area creation after a exec*() on ARM
For some reason, single-stepping over a software interrupt instruction was
failing on ARM after an exec*() call on the target process. Therefore, the
single-step was replaced by a PTRACE_CONT plus a breakpoint inserted after the
instruction, which seems to work fine on both x86 and ARM.

This commits also modifies the struct syscall_data initialization to use C99
initializers.

Signed-off-by: Anderson Lizardo <anderson.lizardo@indt.org.br>