Functracer: remove /proc/PID/maps copying functions
Remove /proc/PID/maps copying functions, given that address ranges are now
inserted directly into the .trace file. This avoids writing a file to disk when
printing traces directly to stdout and avoids issues with libraries that are
loaded during program execution and unloaded before exiting, so that they don't
appear on the /proc/PID/maps that is saved.

Signed-off-by: Anderson Lizardo <anderson.lizardo@indt.org.br>