Functracer: modify instructions for SSOL support
Some instructions do not work well when single-stepped out of place, because
their behavior is location dependent (e.g. they use the PC register value for
calculations, or write to it). Therefore, it is necessary to rewrite these
instructions so that they become "SSOL safe". This consists on replacing the
location dependent register operand (PC, for instance), with other with no
side-effects (e.g. R0). After the instruction is safely single-stepped, the
result is written back to the correct register before the process continues.

This commit adds such support for functracer, but for sake of simplicity, it
only rewrites those instructions that occur in breakpoints inserted by current
functracer plugins. Functracer only inserts breakpoints on the beggining of
functions, therefore there are only a few instructions that appear there.

Ideally, all (or most) instructions should be supported. For this one has to
write code to recognize and rewrite other instructions.

The concept and the branch displacement macro were based on kprobes code from
the Linux kernel. The kernel code, on the other hand, has support for rewriting
almost all instructions on ARM.

Also note that for x86, no instruction rewriting rules were implemented (yet),
because so far none of the current plugins inserted breakpoints at "SSOL
unsafe" instructions.

Signed-off-by: Anderson Lizardo <anderson.lizardo@indt.org.br>
Signed-off-by: Bruna Moreira <bruna.moreira@indt.org.br>