Fix arm/thumb function type detection for symbol table lookup
We want the lowest bit of function addresses to be set to 1 for
thumb functions, as this is used for function type identification
in the rest of functracer code. Unfortunately libbfd always returns
even addresses and does not seem to provide any way to distinguish
arm and thumb functions via its public API. Not being able to
represent some of the format specific data is a known feature:
    http://sourceware.org/binutils/docs/bfd/BFD-information-loss.html

This makes the problem rather ugly, and apparently it can't be
solved without peeking into internal libbfd structs. What makes
it even worse, these internal data structs have changed in
binutils-2.22, more information can be found at:
    http://sources.redhat.com/ml/gdb-patches/2011-03/msg01105.html

This functracer patch was tested with binutils versions 2.20, 2.21
and 2.22